# openapi.yaml
openapi: 3.0.3
info:
  title: ClearPath API
  version: 1.5.0
  description: |
    Shared REST API contract for ClearPath frontend and backend teams.

    ## Changelog — v1.4.2
    - **Fix 5 (Medical ID / Emergency Contacts, Option B):** Mobile dropped the
      on-device SecureStore + QR approach (Option A) after implementation
      issues. Added `GET`/`PUT /user/medical-id` (`blood_type`, `conditions`,
      `allergies`) and `GET`/`POST /user/emergency-contacts` plus
      `PUT`/`DELETE /user/emergency-contacts/{contact_id}`. This data is now
      synced server-side between Mobile and Web, with full-table encryption
      applied at the DB layer. **This reverses the client-side-only stance
      from Fix 2/Fix 3 (D10) for this data** — pending confirmation from
      whoever owns D10 sign-off that DB-layer encryption satisfies the
      original compliance requirement. The `medical-passport` endpoint is
      unaffected and remains D10-compliant (render-token only, no server PDF
      storage).

    ## Changelog — v1.4.1
    - **Fix 1 (Auth Binding):** `GET /user/profile` and `GET /user/settings` upgraded from
      `ApiKeyAuth` to `BearerAuth` (JWT) so the backend can isolate per-user MySQL sessions.
    - **Fix 2 (Medical Data / D10 Compliance):** `GET /user/medical-passport` no longer
      returns a server-hosted `pdf_url`. The response is a lightweight audit acknowledgement
      containing a short-lived `render_token` (signed JWT, 5-min TTL). The Web frontend
...
servers:
...
paths:
...
  /api/v1/user/profile:
    get:
      tags: [User]
      summary: Get the current user profile
      operationId: getUserProfile
      security:
        - BearerAuth: []   # FIX 1: was ApiKeyAuth — JWT required for per-user session isolation
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Invalid or missing bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # FIX 11 — F-13 / B2.4 Account Personalization
    put:
      tags: [User]
      summary: Update editable profile fields
      description: >
        Updates the user's editable profile fields. `user_id`, `email`, and
        `full_name` are registration-locked (read-only) and cannot be changed
        via this endpoint.
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid or missing bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # FIX 1 — Auth Binding: upgraded from ApiKeyAuth to BearerAuth (JWT)
  /api/v1/user/settings:
    get:
      tags: [User]
      summary: Get user settings
      operationId: getUserSettings
      security:
        - BearerAuth: []   # FIX 1: was ApiKeyAuth — JWT required for per-user session isolation
      responses:
        '200':
          description: User settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          description: Invalid or missing bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # FIX 11 — F-13 / B2.4 Account Personalization
    put:
      tags: [User]
      summary: Update user settings
      operationId: updateUserSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsUpdate'
      responses:
        '200':
          description: Settings updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid or missing bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/languages:
    get:
      tags: [User]
      summary: List language options
      operationId: listLanguageOptions
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Supported language options.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguageOptionListResponse'
        '401':
          description: Invalid or missing API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/favourites:
    get:
      tags: [User]
      summary: List saved favourites
      operationId: listFavourites
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Saved venues.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteListResponse'
        '401':
          description: Invalid or missing API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # FIX 4 — Web Read-Only Gating
    post:
      tags: [User]
      summary: Add a venue to favourites
      description: >
        **Mobile-only write endpoint.**
        Returns `403 Forbidden` when `X-Client-Origin: web` is detected.
      x-web-readonly: true
      operationId: addFavourite
      security:
        - BearerAuth: []
      parameters:
        - name: X-Client-Origin
          in: header
          required: false
          description: >
            Client platform identifier. Must be `mobile` for write access.
            Requests with value `web` are rejected with 403 Forbidden.
          schema:
            type: string
            enum: [mobile, web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteCreateRequest'
      responses:
        '201':
          description: Favourite created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favourite'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden. Write operations are not permitted from the Web client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/favourites/{venue_id}:
    delete:
      tags: [User]
      summary: Remove a venue from favourites
      operationId: deleteFavourite
      security:
        - BearerAuth: []
      parameters:
        - name: venue_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Favourite removed.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Favourite not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/notification-preferences:
    get:
      tags: [User]
      summary: Get busyness alert and notification preferences
      operationId: getNotificationPreferences
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification preferences.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [User]
      summary: Update busyness alert and notification preferences
      operationId: updateNotificationPreferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferencesUpdate'
      responses:
        '200':
          description: Notification preferences updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/sos:
    post:
      tags: [User]
      summary: Trigger SOS panic flow
      operationId: triggerSos
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SosRequest'
      responses:
        '200':
          description: SOS event accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SosResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/account:
    delete:
      tags: [User]
      summary: Delete user account and purge synced data
      operationId: deleteUserAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deletion accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAccountResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # FIX 2 & FIX 3 — Medical Data Starvation / Server-Side Storage Risk
  # Per D10, all sensitive medical data is strictly client-side. This endpoint
  # no longer triggers server-side PDF generation or storage. Instead it returns
  # a short-lived render_token (signed JWT, 5-min TTL) that authorises the Web
  # frontend to render the passport client-side after the user scans the mobile
  # QR code. Zero clinical rows or PDF files are written to the cloud.
  /api/v1/user/medical-passport:
    get:
      tags: [User]
      summary: Issue a render token for client-side medical passport generation
      description: >
        **D10 Compliant — No server-side PDF storage.**

        This endpoint acts as a lightweight audit log and token issuer only.
        The Web frontend renders the A4 passport PDF entirely client-side using
        data already held locally on the user's device, accessed after scanning
        the mobile QR code that encodes the `render_token` returned here.

        No clinical payload is transmitted to or stored on the server. The
        `pdf_url` field from v1.4.0 has been removed (Fix 3 — Server-Side
        Storage Risk).
      operationId: getMedicalPassport
      security:
        - BearerAuth: []
      parameters:
        - name: language
          in: query
          required: false
          schema:
            type: string
            example: en
      responses:
        '200':
          description: >-
            Render token issued. No clinical data is stored server-side.
            The client uses the render_token to authorise local PDF rendering
            after QR scan. Token expires in 5 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalPassportResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # FIX 5 — Medical ID / Emergency Contacts (Option B, supersedes D10 client-only mandate)
  # Mobile dropped the on-device SecureStore + QR approach (Option A) after
  # implementation issues; Medical ID and emergency contacts are now synced
  # server-side between Mobile and Web. The DB layer applies full-table
  # encryption to this data; this reverses the client-side-only stance taken
  # for Fix 2/Fix 3 on the medical-passport endpoint above, pending D10 sign-off.
  /api/v1/user/medical-id:
    get:
      tags: [User]
      summary: Get the user's Medical ID
      operationId: getMedicalId
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Medical ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalId'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [User]
      summary: Update the user's Medical ID
      operationId: updateMedicalId
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalIdUpdate'
      responses:
        '200':
          description: Medical ID updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalId'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/emergency-contacts:
    get:
      tags: [User]
      summary: List emergency contacts
      operationId: listEmergencyContacts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Emergency contacts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyContactListResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [User]
      summary: Add an emergency contact
      operationId: addEmergencyContact
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyContactCreateRequest'
      responses:
        '201':
          description: Emergency contact created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyContact'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/emergency-contacts/{contact_id}:
    put:
      tags: [User]
      summary: Update an emergency contact
      operationId: updateEmergencyContact
      security:
        - BearerAuth: []
      parameters:
        - name: contact_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyContactUpdate'
      responses:
        '200':
          description: Emergency contact updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyContact'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Emergency contact not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [User]
      summary: Remove an emergency contact
      operationId: deleteEmergencyContact
      security:
        - BearerAuth: []
      parameters:
        - name: contact_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Emergency contact removed.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Emergency contact not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotificationPreferences:
      type: object
      properties:
        busyness_alerts_enabled:
          type: boolean
        push_notifications_enabled:
          type: boolean
        quiet_hours_enabled:
          type: boolean
        quiet_hours_start:
          type: string
          example: '22:00'
        quiet_hours_end:
          type: string
          example: '07:00'
        alert_threshold_percent:
          type: integer
          minimum: 0
          maximum: 100
        preferred_venue_types:
          type: array
          items:
            type: string
        preferred_boroughs:
          type: array
          items:
            type: string
    NotificationPreferencesUpdate:
      type: object
      additionalProperties: false
      properties:
        busyness_alerts_enabled:
          type: boolean
        push_notifications_enabled:
          type: boolean
        quiet_hours_enabled:
          type: boolean
        quiet_hours_start:
          type: string
        quiet_hours_end:
          type: string
        alert_threshold_percent:
          type: integer
          minimum: 0
          maximum: 100
        preferred_venue_types:
          type: array
          items:
            type: string
        preferred_boroughs:
          type: array
          items:
            type: string
    SosRequest:
      type: object
      additionalProperties: false
      properties:
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        share_live_location:
          type: boolean
        note:
          type: string
    SosResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: accepted
        message:
          type: string
          example: SOS alert triggered successfully.
        incident_id:
          type: string
    DeleteAccountResponse:
      type: object
      required: [status, message, purge_deadline]
      properties:
        status:
          type: string
          example: scheduled
        message:
          type: string
        purge_deadline:
          type: string
          format: date-time

    # FIX 2 & FIX 3 — Medical Data / Server-Side Storage Risk
    # pdf_url removed. No clinical PDF is stored server-side (D10 mandate).
    # render_token is a short-lived signed JWT (5-min TTL) that the Web frontend
    # exchanges after QR scan to authorise purely client-side PDF rendering.
    MedicalPassportResponse:
      type: object
      required: [status]
      description: >
        Medical passport generation acknowledgement. No clinical PDF is stored
        server-side (D10 mandate). The Web frontend renders the passport
        client-side from locally-held data after scanning the mobile QR code.
        This response acts as a lightweight audit log entry only.
      properties:
        status:
          type: string
          example: generated
        render_token:
          type: string
          description: >
            Short-lived signed JWT (5-minute TTL) the Web frontend exchanges
            via QR scan to authorise client-side PDF rendering. Contains no
            clinical payload — only a user reference and expiry claim.
        language:
          type: string
        generated_at:
          type: string
          format: date-time

    # FIX 5 — Medical ID / Emergency Contacts (Option B)
    # Synced server-side between Mobile and Web; DB applies full-table
    # encryption. Supersedes the client-side-only stance from Fix 2/Fix 3
    # for this data, pending D10 sign-off.
    MedicalId:
      type: object
      properties:
        blood_type:
          type: string
          example: O+
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
    MedicalIdUpdate:
      type: object
      additionalProperties: false
      properties:
        blood_type:
          type: string
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ValidationErrorResponse:
      type: object
      required: [error, missing_fields]
      properties:
        error:
          type: string
        missing_fields:
          type: array
          items:
            type: string
        invalid_fields:
          type: array
          items:
            type: string
        allowed_issue_types:
          type: array
          items:
            type: string
        allowed_actions:
          type: array
          items:
            type: string
...
    UserProfile:
      type: object
      required: [user_id, account_state, full_name]
      properties:
        user_id:
          type: string
        account_state:
          type: string
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        date_of_birth:
          type: string
        gender:
          type: string
        nationality:
          type: string
        address:
          type: string
        spoken_languages:
          type: array
          items:
            type: string
        guest_prompt_title:
          type: string
        avatar_initials:
          type: string
    UserSettings:
      type: object
      properties:
        selected_language:
          type: string
        selected_language_native:
          type: string
        location_access_enabled:
          type: boolean
        notifications_enabled:
          type: boolean
        privacy_mode:
          type: string
        guest_mode_enabled:
          type: boolean
        show_medical_id_on_sos:
          type: boolean
        data_export_available:
          type: boolean
        delete_account_enabled:
          type: boolean

    # FIX 11 — F-13 / B2.4 Account Personalization
    # user_id, email, and full_name are registration-locked (read-only) and
    # are intentionally omitted — only these 3 fields are user-editable.
    UserProfileUpdate:
      type: object
      additionalProperties: false
      description: >
        Editable profile fields only. `user_id`, `email`, and `full_name`
        are registration-locked (read-only) and cannot be updated via this
        endpoint.
      properties:
        phone:
          type: string
        nationality:
          type: string
        spoken_languages:
          type: array
          items:
            type: string

    # FIX 11 — F-13 / B2.4 Account Personalization
    UserSettingsUpdate:
      type: object
      additionalProperties: false
      properties:
        selected_language:
          type: string
        selected_language_native:
          type: string
        location_access_enabled:
          type: boolean
        notifications_enabled:
          type: boolean
        privacy_mode:
          type: string
        guest_mode_enabled:
          type: boolean
        show_medical_id_on_sos:
          type: boolean
    LanguageOptionListResponse:
      type: object
      required: [count, items]
      properties:
        count:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/LanguageOption'
...
    FavouriteListResponse:
      type: object
      required: [count, items]
      properties:
        count:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/Favourite'
    Favourite:
      type: object
      required: [favourite_id, venue_id]
      properties:
        favourite_id:
          type: string
        venue_id:
          type: string
        saved_at:
          type: string
          format: date-time
        display_status:
          type: string
    FavouriteCreateRequest:
      type: object
      required: [venue_id]
      properties:
        venue_id:
          type: string
...